addon/spiderbridge/app/requirements.txt
paho-mqtt>=1.6.1,<2.0
pyyaml>=6.0
pytest>=7.4.0
pytest-mock>=3.11
